Update conditional check for optional values in wireframe
This commit updates the conditional check for whether or not a
wireframe's base view controller is already presenting a view
controller when trying to route to the wireframe.

Previously we could check for the presence of an optional value in a
conditional by just calling the optional value. With the upgrade of
xcode beta 6, we have to explicitly check that the value `!=` nil.